fixing various mapping issues to match the schema adding sample user data file

+++ src/main/java/com/sjsu/bikelet/domain/AddressAssociation.java
@@ -11,7 +11,7 @@ public class AddressAssociation {
 
     private Integer entityId;
 
-    private Integer addressId;
+    private Long addressId;
 
     private Integer entityType;
 }
